+++ geopackage-sdk/src/androidTest/java/mil/nga/geopackage/extension/nga/index/FeatureTableIndexImportTest.java
@@ -28,7 +28,20 @@ public class FeatureTableIndexImportTest extends ImportGeoPackageTestCase {
     @Test
     public void testIndex() throws Exception {
 
-        FeatureTableIndexUtils.testIndex(geoPackage);
+        FeatureTableIndexUtils.testIndex(geoPackage, false);
+
+    }
+
+    /**
+     * Test index
+     *
+     * @throws Exception
+     *             upon error
+     */
+    @Test
+    public void testIndexGeodesic() throws Exception {
+
+        FeatureTableIndexUtils.testIndex(geoPackage, true);
 
     }