+++ geopackage-sdk/src/androidTest/java/mil/nga/geopackage/extension/nga/index/FeatureTableIndexUtils.java
@@ -25,6 +25,7 @@ import mil.nga.proj.ProjectionFactory;
 import mil.nga.proj.ProjectionTransform;
 import mil.nga.sf.GeometryEnvelope;
 import mil.nga.sf.Point;
+import mil.nga.sf.proj.ProjectionGeometryUtils;
 import mil.nga.sf.util.GeometryEnvelopeBuilder;
 
 /**
@@ -37,10 +38,11 @@ public class FeatureTableIndexUtils {
     /**
      * Test read
      *
-     * @param geoPackage
+     * @param geoPackage GeoPackage
+     * @param geodesic   index using geodesic bounds
      * @throws Exception
      */
-    public static void testIndex(GeoPackage geoPackage) throws Exception {
+    public static void testIndex(GeoPackage geoPackage, boolean geodesic) throws Exception {
 
         // Test indexing each feature table
         List<String> featureTables = geoPackage.getFeatureTables();
@@ -48,7 +50,7 @@ public class FeatureTableIndexUtils {
 
             FeatureDao featureDao = geoPackage.getFeatureDao(featureTable);
             FeatureTableIndex featureTableIndex = new FeatureTableIndex(
-                    geoPackage, featureDao);
+                    geoPackage, featureDao, geodesic);
 
             // Determine how many features have geometry envelopes or geometries
             int expectedCount = 0;
@@ -107,7 +109,8 @@ public class FeatureTableIndexUtils {
                     .query();
             while (featureTableResults.hasNext()) {
                 GeometryIndex geometryIndex = featureTableResults.next();
-                validateGeometryIndex(featureTableIndex, geometryIndex);
+                validateGeometryIndex(featureTableIndex, geometryIndex,
+                        geodesic);
                 resultCount++;
             }
             featureTableResults.close();
@@ -133,7 +136,8 @@ public class FeatureTableIndexUtils {
             featureTableResults = featureTableIndex.query(envelope);
             while (featureTableResults.hasNext()) {
                 GeometryIndex geometryIndex = featureTableResults.next();
-                validateGeometryIndex(featureTableIndex, geometryIndex);
+                validateGeometryIndex(featureTableIndex, geometryIndex,
+                        geodesic);
                 if (geometryIndex.getGeomId() == testFeatureRow.getId()) {
                     featureFound = true;
                 }
@@ -170,7 +174,8 @@ public class FeatureTableIndexUtils {
                     transformedBoundingBox, projection);
             while (featureTableResults.hasNext()) {
                 GeometryIndex geometryIndex = featureTableResults.next();
-                validateGeometryIndex(featureTableIndex, geometryIndex);
+                validateGeometryIndex(featureTableIndex, geometryIndex,
+                        geodesic);
                 if (geometryIndex.getGeomId() == testFeatureRow.getId()) {
                     featureFound = true;
                 }
@@ -199,7 +204,8 @@ public class FeatureTableIndexUtils {
             featureTableResults = featureTableIndex.query(envelope);
             while (featureTableResults.hasNext()) {
                 GeometryIndex geometryIndex = featureTableResults.next();
-                validateGeometryIndex(featureTableIndex, geometryIndex);
+                validateGeometryIndex(featureTableIndex, geometryIndex,
+                        geodesic);
                 if (geometryIndex.getGeomId() == testFeatureRow.getId()) {
                     featureFound = true;
                 }
@@ -353,13 +359,18 @@ public class FeatureTableIndexUtils {
      * @param geometryIndex
      */
     private static void validateGeometryIndex(
-            FeatureTableIndex featureTableIndex, GeometryIndex geometryIndex) {
+            FeatureTableIndex featureTableIndex, GeometryIndex geometryIndex,
+            boolean geodesic) {
         FeatureRow featureRow = featureTableIndex.getFeatureRow(geometryIndex);
         TestCase.assertNotNull(featureRow);
         TestCase.assertEquals(featureTableIndex.getTableName(),
                 geometryIndex.getTableName());
         TestCase.assertEquals(geometryIndex.getGeomId(), featureRow.getId());
         GeometryEnvelope envelope = featureRow.getGeometryEnvelope();
+        if (geodesic) {
+            envelope = ProjectionGeometryUtils.geodesicEnvelope(envelope,
+                    featureTableIndex.getProjection());
+        }
 
         TestCase.assertNotNull(envelope);