+++ geopackage-sdk/src/androidTest/java/mil/nga/geopackage/features/OAPIFeatureGeneratorTest.java
@@ -40,7 +40,7 @@ public class OAPIFeatureGeneratorTest extends CreateGeoPackageTestCase {
     }
 
     /**
-     * Test flurstueck
+     * Test lakes
      *
      * @throws SQLException upon failure
      */
@@ -82,7 +82,7 @@ public class OAPIFeatureGeneratorTest extends CreateGeoPackageTestCase {
     }
 
     /**
-     * Test mage
+     * Test MAGE
      *
      * @throws SQLException upon failure
      */